requirements.txt
numpy==2.1.1
opencv-python==4.10.0.84
pygame==2.6.0
python-dotenv==1.0.1
